Extract `ValuesYouCanReference` component
