+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/ClickBehaviorSidebar.jsx
@@ -5,13 +5,11 @@ import { getIn } from "icepick";
 import _ from "underscore";
 import cx from "classnames";
 
-import AccordionList from "metabase/core/components/AccordionList";
 import Button from "metabase/core/components/Button";
 import Icon from "metabase/components/Icon";
 import InputBlurChange from "metabase/components/InputBlurChange";
 import ModalContent from "metabase/components/ModalContent";
 import ModalWithTrigger from "metabase/components/ModalWithTrigger";
-import PopoverWithTrigger from "metabase/components/PopoverWithTrigger";
 
 import { color } from "metabase/lib/colors";
 import {
@@ -31,11 +29,10 @@ import QuestionPicker from "metabase/containers/QuestionPicker";
 
 import Sidebar from "metabase/dashboard/components/Sidebar";
 import ClickMappings, {
-  withUserAttributes,
   clickTargetObjectType,
-  isMappableColumn,
 } from "metabase/dashboard/components/ClickMappings";
 
+import ValuesYouCanReference from "./ValuesYouCanReference";
 import {
   SidebarItemWrapper,
   SidebarItemClasses,
@@ -847,60 +844,5 @@ const CustomLinkText = ({ clickBehavior, updateSettings }) => {
     </div>
   );
 };
-
-const ValuesYouCanReference = withUserAttributes(
-  ({ dashcard, parameters, userAttributes }) => {
-    const columnMetadata = dashcard.card.result_metadata || [];
-    const columns = columnMetadata?.filter(isMappableColumn).map(c => c.name);
-    const parameterNames = parameters.map(p => p.name);
-    const sections = [
-      {
-        items: prefixIfNeeded(columns, "column", [
-          parameterNames,
-          userAttributes,
-        ]),
-        name: t`Columns`,
-      },
-      {
-        items: prefixIfNeeded(parameterNames, "filter", [
-          columns,
-          userAttributes,
-        ]),
-        name: t`Dashboard filters`,
-      },
-      {
-        items: prefixIfNeeded(userAttributes, "user", [
-          parameterNames,
-          columns,
-        ]),
-        name: t`User attributes`,
-      },
-    ].filter(section => section.items.length > 0);
-    return (
-      <PopoverWithTrigger
-        triggerElement={
-          <div className="flex align-center cursor-pointer my2 text-medium text-brand-hover">
-            <h4>{t`Values you can reference`}</h4>
-            <Icon name="chevrondown" className="ml1" size={12} />
-          </div>
-        }
-      >
-        <AccordionList
-          alwaysExpanded
-          sections={sections}
-          renderItemName={name => name}
-          itemIsClickable={() => false}
-        />
-      </PopoverWithTrigger>
-    );
-  },
-);
-
-function prefixIfNeeded(values, prefix, otherLists) {
-  const otherValues = otherLists.flat().map(s => s.toLowerCase());
-  return values.map(value =>
-    otherValues.includes(value.toLowerCase()) ? `${prefix}:${value}` : value,
-  );
-}
 
 export default ClickBehaviorSidebar;
